+++ app/(routes)/project-gallery/Gallery.module.css
@@ -193,20 +193,6 @@
   overflow: hidden;
 }
 
-.clickableImage {
-  cursor: pointer;
-  transition: transform 0.4s ease;
-}
-
-.featuredCard:hover .clickableImage {
-  background: linear-gradient(
-    to top,
-    rgba(0, 0, 0, 0.9) 0%,
-    rgba(0, 0, 0, 0.5) 70%,
-    transparent 100%
-  );
-}
-
 .featuredOverlay {
   position: absolute;
   bottom: 0;
@@ -280,27 +266,6 @@
   position: relative;
   height: 250px;
   overflow: hidden;
-}
-
-.placeholderImage {
-  width: 100%;
-  height: 100%;
-  background: #f0f0f0;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
-  color: #999;
-  border: 1px dashed #ccc;
-}
-
-.placeholderImage p {
-  margin-top: 0.5rem;
-  font-size: 0.9rem;
-}
-
-.projectCard:hover .projectImage .clickableImage {
-  transform: scale(1.05);
 }
 
 .imageCount {
@@ -726,126 +691,13 @@
   }
 }
 
-/* Carousel Controls */
-.carouselControl {
-  position: absolute;
-  top: 50%;
-  transform: translateY(-50%);
-  background: rgba(0, 0, 0, 0.6);
-  color: white;
-  border: none;
-  width: 40px;
-  height: 40px;
-  border-radius: 50%;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  cursor: pointer;
-  z-index: 10;
-  opacity: 0;
-  transition: all 0.3s ease;
-  font-size: 1.2rem;
-}
-
-.featuredCard:hover .carouselControl {
-  opacity: 1;
-}
-
-.carouselControl:hover {
-  background: linear-gradient(
-    to top,
-    rgba(0, 0, 0, 0.9) 0%,
-    rgba(0, 0, 0, 0.5) 70%
-  );
-}
-
-.carouselPrev {
-  left: 15px;
-}
-
-.carouselNext {
-  right: 15px;
-}
-
-/* Carousel Indicators */
-.carouselIndicators {
-  position: absolute;
-  bottom: 80px; /* Above the overlay */
-  left: 50%;
-  transform: translateX(-50%);
-  display: flex;
-  gap: 8px;
-  z-index: 10;
-}
-
-.carouselIndicator {
-  width: 8px;
-  height: 8px;
-  border-radius: 50%;
-  border: none;
-  background: rgba(255, 255, 255, 0.5);
-  cursor: pointer;
-  transition: all 0.3s ease;
-}
-
-.carouselIndicator.active,
-.carouselIndicator:hover {
-  background: white;
-  transform: scale(1.2);
-}
-
-/* Image Counter */
-.imageCounter {
-  position: absolute;
-  top: 15px;
-  right: 15px;
-  background: rgba(0, 0, 0, 0.7);
-  color: white;
-  padding: 4px 8px;
-  border-radius: 12px;
-  font-size: 0.8rem;
-  font-weight: 500;
-  z-index: 10;
-}
-
 /* Featured Image positioning */
 .featuredImage {
   position: relative;
   overflow: hidden;
 }
 
-/* Carousel Wrapper */
-.carouselWrapper {
-  position: relative;
-  width: 100%;
-  height: 100%;
-  overflow: hidden;
-}
 
-/* Project card specific carousel styling */
-.projectImageCarousel .carouselControl {
-  opacity: 0;
-  transition: opacity 0.2s ease;
-}
-
-.projectCard:hover .carouselControl {
-  opacity: 1;
-}
-
-/* Smaller controls for project cards */
-.projectImageCarousel .carouselControl {
-  width: 32px;
-  height: 32px;
-  font-size: 1rem;
-}
-
-.projectImageCarousel .carouselPrev {
-  left: 8px;
-}
-
-.projectImageCarousel .carouselNext {
-  right: 8px;
-}
 
 /* Updated Project Modal Structure */
 .projectModal {
